+++ Source/Lutefisk3D/Core/WorkQueue.h
@@ -114,7 +114,7 @@ public:
     void SetNonThreadedWorkMs(int ms) { maxNonThreadedWorkMs_ = std::max(ms, 1); }
 
     /// Return number of worker threads.
-    unsigned GetNumThreads() const { return threads_.size(); }
+    size_t GetNumThreads() const { return threads_.size(); }
     /// Return whether all work with at least the specified priority is finished.
     bool IsCompleted(unsigned priority) const;
     /// Return whether the queue is currently completing work in the main thread.
